Fix AVX2 bugs
